==> 02-radiation_response.Rmd <==
# In vitro radiation response and clinical data

## Progression-free survival



```{r libraries and sources for progression-free survival analysis, echo = T, message=FALSE, warning=FALSE}
# Libraries and sources
library(readxl)
library(tidyverse)
library(ggplot2)
library(ggfortify)
library(survival) 
library(survminer) 
library(survMisc)
library(cowplot)
```

```{r wrangle for Progression-free survival, echo = T, message=FALSE, warning=FALSE}
########
# Data #
########
# Loading survival data from time of surgery
surg_surv <- as.data.frame(read_excel("../cna_analysis/data/clinical_and_dr/data_survival_progression_with_death_as_progression.xlsx", 
                      sheet = "surg_surv"))

# Loading progression data from time of surgery
surg_prog <- as.data.frame(read_excel("../cna_analysis/data/clinical_and_dr/data_survival_progression_with_death_as_progression.xlsx", 
                        sheet = "surg_prog"))

#######
# Run #
#######
## Survival from time of surgery ##
surg_surv_sum <- Surv(time = surg_surv$time, event = surg_surv$status == 2)

# Fit without strata (~1)
surg_surv_fit <- survfit(surg_surv_sum ~ 1)

# Median survival from time of surgery
surv_median(surg_surv_fit)[1,2]

## Progression from time of surgery ##
surg_prog_sum <- Surv(time = surg_prog$time, event = surg_prog$status == 2)

# Fit without strata (~1)
surg_prog_fit <- survfit(surg_prog_sum ~ 1)

# Median progression from time of surgery
surv_median(surg_prog_fit)[1,2]

```

```{r plot Progression-free survival, echo = T, message=FALSE, warning=FALSE}
#### PROGRESSION ####
# Constructing a dataframe for annotating labels
anno_df_prog <- data.frame(label = surg_prog$organoid, 
                           time = surg_prog$time)

# sort by time
anno_df_prog <- anno_df_prog[order(anno_df_prog$time),]

# make temp df with survival/prob data
temp_df_prob_prog <- data.frame(time = surg_prog_fit$time,
                                prob = surg_prog_fit$surv
                                )
# join together
anno_df_prog <- left_join(anno_df_prog, temp_df_prob_prog)



# Plotting progression free survival
prog_p <- ggsurvplot(surg_prog_fit, data = surg_prog, conf.int = F, palette = c("#00887d"),
          ggtheme = theme_cowplot(),
           legend = 'none',
           censor.size = 8,
           xlim = c(0,2200),
          ylab = 'Progression-free survival probability', 
          xlab = 'Time from surgery (days)')

# adding labels
prog_p$plot <- prog_p$plot + 
  annotate("text", 
                x = anno_df_prog$time + c(350, 350, 350, -150,50,250,350,0), 
           y = anno_df_prog$prob + c(0.1,0.1,0.1, -0.05, 0.05, -0.05, 0.28,-0.05),
                label = anno_df_prog$label) + 
  theme(aspect.ratio = 5/3) + 
  coord_cartesian(clip = 'off')

prog_p 
```

## In vitro dose-response 
```{r libraries and sources for in vitro dose-response, echo = T, message=FALSE, warning=FALSE}
# Libraries and sources
library(tidyverse) # for wrangling data
library(readxl) # for reading excel files
library(DescTools) # for computing AUCs with AUC() function)
library(drc) # for fitting dose response 
library(cowplot)
library(wesanderson)
source('R/auc_rad.R')
source('R/fitter_rad.R')
source('R/ic50_rad.R')
source('R/relativize_dr.R')
```
### Dose-response plot

```{r relativize for in vitro dose-response, echo = T, message=FALSE, warning=FALSE}
################
# Loading data #
################
d <- data.frame(read_excel('../cna_analysis/data/clinical_and_dr/resistancy_organoids.xlsx'))

# remove fluke experiment (can't be fitted by the drc package in metricsandfits.R)
d <- subset(d, expcode != 'cva0028')

# remove dose 1, 10,12
d <- subset(d, dose %in% c(0,2,4,6,8))

# only include organoids that were not radiated
d <- subset(d, rec_rad == 0)

##############
# Relativize #
##############
# Relativizing the data

# split per experiment
per_exp <- split(d, f = d$expcode, drop = TRUE)

# apply relative function to every instance in per_exp
rel_vector_unrad <- lapply(per_exp, relative, rad = F)

# join the dataframes togheter
d_nonrad = do.call(rbind, rel_vector_unrad)

#remove rownames
rownames(d_nonrad) <- NULL

# pool d_rad and d_nonrad, if not possible, make d d_nonrad

# try to bind d_rad and d_nonrad. If d_rad is empty (this is the case when
# rec_rad does not have ones, make d the same as d_nonrad)
d <- d_nonrad
```

```{r fitting and computing metrics for in vitro dose-response, echo = T, message=FALSE, warning=FALSE}
################
# Loading data #
################
load("../cna_analysis/rda/clinical_and_dr/wrangled-data.rda") # load the wrangled data, data is saved under 'd'

# defining experiment number for reference
exp_numb <- unique(d$expcode)
exp_numb <- paste(unlist(exp_numb), collapse='')

##################
### Metric AUC ###
##################
# split per experiment
per_exp <- split(d, f = d$expcode, drop = TRUE)
# apply relative function to every instance in per_exp
AUC_vector <- lapply(per_exp, AUC_fun)

# join the dataframes togheter
AUC = do.call(rbind, AUC_vector)

# remove rownames
rownames(AUC) <- NULL

#######
# Fit #
#######
# apply fit function to every instance in per_exp
fit_vector <- lapply(per_exp, fitter)

# join the dataframes toghether
fit = do.call(rbind, fit_vector)

# remove rownames
rownames(fit) <- NULL

fit$rec_rad = 0 

# remove rownames
rownames(fit) <- NULL

##################
# relative IC50s #
##################
## relative under curve for unrad
# apply relative function to every instance in per_exp
IC50_vector <- lapply(per_exp, IC50_fun, rad = 0)

# join the dataframes togheter
IC50= do.call(rbind, IC50_vector)

metrics = cbind(AUC, IC50)

metrics$rec_rad <- 0

# compute mean of rel_AUC
metrics_pooled <- metrics %>% group_by(cline) %>% summarise_at(vars(rel_AUC),
                                                               funs(mean(., na.rm=TRUE)))


metrics_pooled$expcode <- exp_numb

# compute mean of rel_AUC
metrics_pooled <- metrics %>% group_by(cline) %>% summarise_at(vars(rel_AUC),
                                                               funs(mean(., na.rm=TRUE)))


# summary statistics
median(metrics$rel_AUC)
min(metrics$rel_AUC)
max(metrics$rel_AUC)

```

```{r plots for in vitro dose-response, echo = T, message=FALSE, warning=FALSE}
########
# Data #
########
load("../cna_analysis/rda/clinical_and_dr/wrangled-data.rda") # load the wrangled data, data is saved under 'd'
load("../cna_analysis/rda/clinical_and_dr/fitted-data.rda") # load the wrangled data, data is saved under 'fit'
load('../cna_analysis/rda/clinical_and_dr/metrics_data.rda') # load the metrics data

###########
# Wrangle #
###########
### define two dataframes, one from rad one without rad:

fit <- subset(fit, rec_rad == 0)

exp_numb <- unique(d$expcode)
exp_numb <- paste(unlist(exp_numb), collapse='')

# new data with doses. Note: lowest dose is not
# log 0 but log('very small number') because otherwise this will hamper the scaling in ggplot later on.
d$dose0 <- d$dose
d$dose0[d$dose0 == 0] <- 0.00000001

# representative expcodes
# makeing helper column to extract representative expcodes
d$cline_expcode <- paste0(d$cline, '_', d$expcode)
fit$cline_expcode <- paste0(fit$cline, '_', fit$expcode)

# subsetting
cline_expcode_pick <-
  c(
    'HUB005_cva0025',
    'HUB183_cva0025',
    'HUB015_cva0023',
    'HUB175_cva0024',
    'HUB181I_cva0021',
    'HUB197_cva0030',
    'HUB106_cva0020',
    'HUB062_cva0023',
    'HUB112_cva0112')
    
d <- subset(d, cline_expcode %in% cline_expcode_pick)

fit <- subset(fit, cline_expcode %in% cline_expcode_pick)

# add resistantcy group for colour coding
resistant <- c('HUB005', 'HUB183', 'HUB015')
sensitive <- c('HUB197', 'HUB106', 'HUB062')

d$resistancy_group <-
  ifelse(
    d$cline %in% resistant,
    'Resistant',
    ifelse(d$cline %in% sensitive, 'Sensitive', 'Intermediate')
  )

fit$resistancy_group <-
  ifelse(
    fit$cline %in% resistant,
    'Resistant',
    ifelse(fit$cline %in% sensitive, 'Sensitive', 'Intermediate')
  )

# Make cline a factor and specify order by decreasing AUC mean
d$cline <- factor(d$cline, levels = c('HUB005', 'HUB183', 'HUB015', 
                                      'HUB112', 'HUB175', 'HUB181I', 
                                      'HUB106', 'HUB197', 'HUB062'))

d$resistancy_group <- factor(d$resistancy_group, levels = c('Resistant', 'Intermediate', 'Sensitive'))

ggplot(NULL, aes(x = dose0, y = relative_mean, group = cline, col = resistancy_group)) +
  geom_point(data= d,
             size =3, aes(shape = cline)) +
  geom_errorbar(data = d,
                aes(ymin=relative_mean-relative_sem,ymax=relative_mean+relative_sem),
                width=0.1) +
  geom_line(data = d) + 
  # geom_line(data=fit, aes(x=dose, y=pred),
  #           size = 1) + 
  # adding log x-axis
  coord_trans(#x="log10", 
    xlim = c(0.00000001, 8), ylim = c(0, 1.1)) +
  scale_x_continuous(breaks = c(0, 2,4,6,8)) + 
  scale_y_continuous(expand = c(0,0)) + 
  theme_cowplot() + 
  ylab('Normalized viability') + 
  xlab('Dose (Gy)') + 
  theme(
    legend.title = element_blank(),
    aspect.ratio = 5 / 3) + 
  scale_shape_manual(values=seq(0,8)) + 
  scale_colour_manual(
    values = c(
      Resistant = '#DE3163',
      Sensitive = '#5090CD',
      Intermediate = 'grey75'
    ),
    guide = guide_legend(override.aes = list(
      linetype = c(rep("blank", length(
        unique(d$resistancy_group)
      ))),
      shape = c(rep(19, length(
        unique(d$resistancy_group)
      )))
    ))
  ) 

```

### rAUC across multiple experiments
```{r plot metrics plot in vitro dose-response, echo = T, message=FALSE, warning=FALSE}
########
# Data #
########
load('../cna_analysis/rda/clinical_and_dr/metrics_data.rda') # load the wrangled data, data is saved under 'metrics'

exp_numb <- unique(metrics$expcode)
exp_numb <- paste(unlist(exp_numb), collapse = '')

###########
# Wrangle #
###########
# Make cline a factor and specify order by decreasing AUC mean
metrics$cline <- factor(
  metrics$cline,
  levels = metrics %>% group_by(cline) %>% summarize(mean_auc = mean(rel_AUC)) %>% arrange(-mean_auc) %>% pull(cline)
)

# add resistantcy group for colour coding
resistant <- c('HUB005', 'HUB183', 'HUB015')
sensitive <- c('HUB197', 'HUB106', 'HUB062')

metrics$resistancy_group <-
  ifelse(
    metrics$cline %in% resistant,
    'Resistant',
    ifelse(metrics$cline %in% sensitive, 'Sensitive', 'Intermediate')
  )

# changing order of resistancy_group
metrics$resistancy_group <- factor(metrics$resistancy_group, levels = c("Resistant", "Intermediate", "Sensitive"))

########
# Plot #
########

ggplot(metrics, aes(x = cline, y = rel_AUC, group = cline)) +
  geom_boxplot(aes(fill = resistancy_group)) +
  geom_jitter(height = 0,
              width = 0.1,
              size = 2) +
  theme_cowplot() +
  scale_fill_manual(
    values = c(
      Resistant = '#DE3163',
      Sensitive = '#5090CD',
      Intermediate = 'grey75'
    ),
    guide = guide_legend(override.aes = list(
      linetype = c(rep("blank", length(
        unique(metrics$resistancy_group)
      ))),
      shape = c(rep(19, length(
        unique(metrics$resistancy_group)
      )))
    ))
  ) +
  xlab('') +
  ylab('Relative AUC') +
  theme(
    legend.title = element_blank(),
    aspect.ratio = 5 / 3,
    axis.text.x = element_text(angle = 45, hjust = 1)
  ) +
  coord_cartesian(clip = 'off') + 
  scale_y_continuous(limits = c(0,1))

# statistical test to compare top three radioresistant and radiosensitive organoids
# note that in the paper, the p-value was added manually. 
# Make cline a factor and specify order by decreasing AUC mean
metrics$cline <- factor(
  metrics$cline,
  levels = metrics %>% group_by(cline) %>% summarize(mean_auc = mean(rel_AUC)) %>% arrange(-mean_auc) %>% pull(cline)
)

# add resistancy group for colour coding
resistant <- c('HUB005', 'HUB183', 'HUB015')
sensitive <- c('HUB197', 'HUB106', 'HUB062')

metrics$resistancy_group <-
  ifelse(
    metrics$cline %in% resistant,
    'Resistant',
    ifelse(metrics$cline %in% sensitive, 'Sensitive', 'Intermediate')
  )

# changing order of resistancy_group
metrics$resistancy_group <- factor(metrics$resistancy_group, levels = c("Resistant", "Intermediate", "Sensitive"))

#######
# Run #
#######
# t.test (by just taking the average of each biological replicate). 
t_df <- metrics %>% group_by(cline) %>% summarize(rAUC= mean(rel_AUC), stdev = sd(rel_AUC))

t_df$resistancy_group <-
  ifelse(
    t_df$cline %in% resistant,
    'Resistant',
    ifelse(t_df$cline %in% sensitive, 'Sensitive', 'Intermediate')
  )

# Highly significant p-value. 
t.test(subset(t_df, resistancy_group == 'Resistant')$rAUC, subset(t_df, resistancy_group == 'Sensitive')$rAUC, paired = F)


```

### Recovery statistics after 10 Gy radiation
```{r recovery statistics after 10 Gy radiation, echo = T, message=FALSE, warning=FALSE}
########
# Data #
########
df <- as.data.frame(read_excel("../cna_analysis/data/clinical_and_dr/ 10gy_recovery_data.xlsx"))

###########
# Wrangle #
###########
df_res <- subset(df, resistancy_group == 'Resistant')
df_sens <- subset(df, resistancy_group == 'Sensitive')

##############
# Statistics #
##############
# Resistant
mean_days_res <- mean(as.numeric(df_res$days_to_full_recovery), na.rm = T)
mean_days_res
sd_days_res <- sd(as.numeric(df_res$days_to_full_recovery), na.rm = T)
sd_days_res

# Sensitive
mean_days_sens <- mean(as.numeric(df_sens$days_to_full_recovery), na.rm = T)
mean_days_sens
sd_days_sens <- sd(as.numeric(df_sens$days_to_full_recovery), na.rm = T)
sd_days_sens
```

### Persisters versus shifters dose-response plots
```{r persisters vs shifters, echo = T, message=FALSE, warning=FALSE}
########
# Data #
########
d <- as.data.frame(read_excel("../cna_analysis/data/clinical_and_dr/parental_vs_radiated.xlsx"))

##############
# Relativize #
##############
# split per experiment
per_exp <- split(d, f = d$expcode, drop = TRUE)

# for radiation
rel_vector_unrad <- lapply(per_exp, relative, rad = F)


# join the dataframes togheter
d_nonrad = do.call(rbind, rel_vector_unrad)

# for radiation
rel_vector_rad <- lapply(per_exp, relative, rad = T)

# for radation
d_rad = do.call(rbind, rel_vector_rad)

# try to bind d_rad and d_nonrad. If d_rad is empty (this is the case when
# rec_rad does not have ones, make d the same as d_nonrad)
d <- try(rbind(d_rad, d_nonrad))
if("try-error" %in% class(d)) d <- d_nonrad

###########
# Wrangle #
###########
# making helper column to extract representative expcodes
d$cline_expcode <- paste0(d$cline, '_', d$expcode, '_', d$rec_rad)

# subsetting
cline_expcode_pick <-
  c('HUB005_cva0026_0',
    'HUB005_cva0026_1',
    'HUB183_cva0025_0',
    'HUB183_cva0025_1',
    'HUB106_cva0028_0',
    'HUB106_cva0028_1',
    'HUB197_cva0117_0',
    'HUB197_cva0117_1',
    'HUB062_cva0061_a_0',
    'HUB062_cva0061_b_1')

d <- subset(d, cline_expcode %in% cline_expcode_pick)

d$persister_group <-
  ifelse(
    d$cline %in% c('HUB005', 'HUB183', 'HUB106'),
    'Persister', 'Shifter')

d$rad <- 
  ifelse(
    d$rec_rad == 0, 
    'Parental', 'Recurrence'
  )

# new data with doses. Note: lowest dose is not
# log 0 but log('very small number') because otherwise this will hamper the scaling in ggplot later on.
d$dose0 <- d$dose
d$dose0[d$dose0 == 0] <- 0.00000001

###########
# Wrangle #
###########
s <- subset(d, persister_group == 'Shifter')

ggplot(NULL, aes(x = dose0, y = relative_mean, group = interaction(cline,rad), col = rad)) +
  geom_point(data= s,
             size =3, aes(shape = cline)) +
  geom_errorbar(data = s,
                aes(ymin=relative_mean-relative_sem,ymax=relative_mean+relative_sem),
                width=0.1) +
  geom_line(data = s) +
  coord_trans(#x="log10", 
    xlim = c(0.00000001, 8), ylim = c(0, 1.2)) +
  scale_x_continuous(breaks = c(0, 2,4,6,8)) + 
  scale_y_continuous(expand = c(0,0), breaks = c(0,0.2, 0.4,0.6,0.8,1.0,1.2)) + 
  theme_cowplot() + 
  ylab('Normalized viability') + 
  xlab('Dose (Gy)') + 
  theme(
    legend.title = element_blank(),
    aspect.ratio = 5 / 3) + 
  scale_shape_manual(values=seq(0,8)) + 
  labs(title = 'Shifters') + 
  scale_colour_manual(
    values = c(
      Parental = '#999999',
      Recurrence = "#E69F00"
    ),
    guide = guide_legend(override.aes = list(
      linetype = c(rep("blank", length(
        unique(d$rad)
      ))),
      shape = c(rep(19, length(
        unique(d$rad)
      )))
    ))
  ) 

p <- subset(d, persister_group == 'Persister')

ggplot(NULL, aes(x = dose0, y = relative_mean, group = interaction(cline,rad), col = rad)) +
  geom_point(data= p,
             size =3, aes(shape = cline)) +
  geom_errorbar(data = p,
                aes(ymin=relative_mean-relative_sem,ymax=relative_mean+relative_sem),
                width=0.1) +
  geom_line(data = p) +
  coord_trans(#x="log10", 
    xlim = c(0.00000001, 8), ylim = c(0, 1.2)) +
  scale_x_continuous(breaks = c(0, 2,4,6,8)) + 
  scale_y_continuous(expand = c(0,0), breaks = c(0,0.2, 0.4,0.6,0.8,1.0,1.2)) + 
  theme_cowplot() + 
  ylab('Normalized viability') + 
  xlab('Dose (Gy)') + 
  theme(
    legend.title = element_blank(),
    aspect.ratio = 5 / 3) + 
  scale_shape_manual(values=seq(0,8)) + 
  labs(title = 'Persisters') + 
  scale_colour_manual(
    values = c(
      Parental = '#999999',
      Recurrence = "#E69F00"
    ),
    guide = guide_legend(override.aes = list(
      linetype = c(rep("blank", length(
        unique(d$rad)
      ))),
      shape = c(rep(19, length(
        unique(d$rad)
      )))
    ))
  ) 

```
